requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.4
iniconfig==1.1.1
lxml==4.6.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-allure-adaptor2==1.8.2
pytest-html==2.1.1
pytest-metadata==1.10.0
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
wheel == 0.35.1